refactor: extract shelf and local-filesystem subsystems into mixins
Fourteen PdsFile methods move out of pdsfile.py into two private mixin modules,
bodies byte-for-byte unchanged:

  _shelves.py   _ShelfMixin    shelf_path_and_lskip, shelf_path_and_key,
                               _get_shelf, _close_shelf, close_all_shelves,
                               shelf_lookup, shelf_path_and_key_for_abspath,
                               info_shelf_expected, shelf_exists_if_expected
  _local_fs.py  _LocalFsMixin  _non_checksum_abspath, os_path_exists,
                               os_path_isdir, os_listdir, glob_glob

The class PdsFile statement stays in pdsfile.py and takes the two mixins as
bases, listed alphabetically. The shelf cache class attributes stay defined on
PdsFile, and the module constant PATH_EXISTS_CACHE_SIZE moves with its
lru_cache consumer and is re-exported.

bisect, fnmatch, functools, pickle and _clean_glob are no longer referenced in
pdsfile.py but stay bound, in the explicit re-export form, so that no name
reachable as pdsfile.pdsfile.<name> is lost.

The ruff ratchet entries for the two new modules are a split of pdsfile.py's:
every code they list was already forgiven for these same lines, and the four
codes that no longer occur in pdsfile.py are dropped from its entry.
