+++ src/pdsfile/pdsfile.py
@@ -3,21 +3,22 @@
 # General pdsfile package & PdsFile class
 ##########################################################################################
 
-import bisect
 import datetime
-import fnmatch
-import functools
-# Nothing below references glob or math, but both are frozen members of this
-# module's public surface (tests/api/api_manifest.json), so it keeps them bound.
-# The redundant `as` alias is the explicit re-export form.
-import glob as glob
-import math as math
 import numbers
 import os
-import pickle
 import PIL
 import re
 import time
+
+# Nothing below references these six, but all of them are frozen members of this
+# module's public surface (tests/api/api_manifest.json), so it keeps them bound.
+# The redundant `as` alias is the explicit re-export form.
+import bisect as bisect
+import fnmatch as fnmatch
+import functools as functools
+import glob as glob
+import math as math
+import pickle as pickle
 
 import pdslogger
 import pdstable
@@ -42,7 +43,6 @@ from .preload_and_cache import (cache_lifetime_for_class,
 # The path helpers live in a private module. Importing them here is also what
 # keeps pdsfile.pdsfile.<name> resolving for callers and for the API freeze.
 from ._path_utils import (_clean_abspath,
-                          _clean_glob,
                           _clean_join,
                           _needs_glob,
                           abspath_for_logical_path,
@@ -51,23 +51,31 @@ from ._path_utils import (_clean_abspath,
                           logical_path_from_abspath,
                           repair_case)
 
-# Re-exported only; nothing below references these. FILE_BYTE_UNITS and
-# selected_path_from_path are frozen members of this module's public surface;
-# _GLOB_CACHE_SIZE is private, and is carried so that no name reachable as
-# pdsfile.pdsfile.<name> is lost. The redundant `as` alias is the explicit
-# re-export form, so they do not read as unused imports.
+# Two groups of PdsFile methods live in private modules as mixins, and the class
+# statement below takes them as bases. The bases are listed alphabetically: the
+# mixins share no attribute name and none shadows a name PdsFile defines itself
+# (tests/api/test_mixin_collisions.py asserts both), so the order carries no
+# meaning and a mechanical rule keeps it checkable as more mixins arrive.
+from ._local_fs import _LocalFsMixin
+from ._shelves import _ShelfMixin
+
+# Re-exported only; nothing below references these. FILE_BYTE_UNITS,
+# PATH_EXISTS_CACHE_SIZE and selected_path_from_path are frozen members of this
+# module's public surface; _GLOB_CACHE_SIZE and _clean_glob are private, and are
+# carried so that no name reachable as pdsfile.pdsfile.<name> is lost. The
+# redundant `as` alias is the explicit re-export form, so they do not read as
+# unused imports.
+from ._local_fs import PATH_EXISTS_CACHE_SIZE as PATH_EXISTS_CACHE_SIZE
 from ._path_utils import (FILE_BYTE_UNITS as FILE_BYTE_UNITS,
                           _GLOB_CACHE_SIZE as _GLOB_CACHE_SIZE,
+                          _clean_glob as _clean_glob,
                           selected_path_from_path as selected_path_from_path)
-
-# Configuration
-PATH_EXISTS_CACHE_SIZE = 200
 
 ##########################################################################################
 # PdsFile class
 ##########################################################################################
 
-class PdsFile(object):
+class PdsFile(_LocalFsMixin, _ShelfMixin, object):
 
     # Configuration
     VOLTYPES = ['volumes', 'calibrated', 'diagrams', 'metadata', 'previews',
@@ -1071,415 +1079,6 @@ class PdsFile(object):
         else:
             return ('PdsFile.' + type(self).__name__ + '("' +
                     self.abspath + '")')
-
-    ############################################################################
-    # Local implementations of basic filesystem operations
-    ############################################################################
-
-    @classmethod
-    def _non_checksum_abspath(cls, abspath):
-        """Return the non-checksum path associated with this checksum file. If the given
-        absolute path does not point to a checksum file, it returns None.
-
-        Keyword arguments:
-            abspath -- the absolute path of the checksum file.
-        """
-
-        # Checksum files need special handling
-        if f'/{cls.PDS_HOLDINGS}/checksums-' in abspath:
-            testpath = abspath.replace('/checksums-', '/')
-
-            for voltype in cls.VOLTYPES:
-                testpath = testpath.replace('_' + voltype + '_md5.txt', '')
-
-            return testpath
-
-        else:
-            return None
-
-    @classmethod
-    @functools.lru_cache(maxsize=PATH_EXISTS_CACHE_SIZE)
-    def os_path_exists(cls, abspath, force_case_sensitive=False):
-        """Return True if the given absolute path points to a file that exists; Return
-        False otherwise. This replaces os.path.exists(path) but might use infoshelf
-        files rather than refer to the holdings directory.
-
-        Note: This function is case-insensitive under SHELVES_ONLY. Otherwise,
-        its behavior matches that of the file system. For Macs, this usually
-        means that it is case insensitive. If force_case_sensitive=True, then
-        the check of the basename will be case-sensitive regardless of the file
-        system.
-
-        Keyword arguments:
-            abspath              -- the absolute path of the file.
-            force_case_sensitive -- a flag to determine if the basename will be case
-                                    sensitive (default False)
-        """
-
-        if f'{cls.PDS_HOLDINGS}/_infoshelf' in abspath:
-            return os.path.exists(abspath)
-
-        # Handle index rows
-        for ext in cls.IDX_EXT:
-            if f'{ext}/' in abspath:
-                parts = abspath.partition(f'{ext}/')
-                if not cls.os_path_exists(parts[0] + ext):
-                    return False
-                pdsf = cls.from_abspath(parts[0] + ext)
-                return (pdsf.exists and
-                        pdsf.child_of_index(parts[2], flag='').exists)
-
-        # If it's for documentation, we don't create shelf files, we will just use the
-        # os.path.exists
-        if cls.SHELVES_ONLY and f'{cls.PDS_HOLDINGS}/documents' not in abspath:
-            try:
-                (shelf_abspath,
-                 key) = cls.shelf_path_and_key_for_abspath(abspath, 'info')
-
-                if key:
-                    shelf = cls._get_shelf(shelf_abspath,
-                                               log_missing_file=False)
-                    return (key in shelf)
-                elif cls.os_path_exists(shelf_abspath):
-                    return True     # Every shelf file has an entry with an
-                                    # empty key, so this avoids an unnecessary
-                                    # open of the file.
-                else:
-                    return False
-            except (ValueError, IndexError, IOError, OSError):
-                pass
-
-            # Maybe it's associated with something else in the infoshelf tree
-            if f'/{cls.PDS_HOLDINGS}/' in abspath:
-
-                # Maybe there's an associated directory in the infoshelf tree
-                shelf_abspath = abspath.replace(f'/{cls.PDS_HOLDINGS}/',
-                                                f'/{cls.PDS_HOLDINGS}/_infoshelf-')
-                if cls.os_path_exists(shelf_abspath):
-                    return True
-
-                # Maybe there's an associated shelf file in the infoshelf tree
-                if cls.os_path_exists(shelf_abspath + '_info.pickle'):
-                    return True
-
-                # Checksum files need special handling, before doing special handling,
-                testpath = cls._non_checksum_abspath(abspath)
-                if testpath and cls.os_path_exists(testpath):
-                    return True
-
-        if force_case_sensitive and cls.FS_IS_CASE_INSENSITIVE:
-            test = os.path.exists(abspath)
-            if not test:
-                return False
-
-            (parent,basename) = os.path.split(abspath)
-            childnames = os.listdir(parent)
-            return (basename in childnames)
-
-        return os.path.exists(abspath)
-
-    @classmethod
-    def os_path_isdir(cls, abspath):
-        """Return True if the given absolute path points to a directory; Return False
-        otherwise. This replaces os.path.isdir() but might use infoshelf files rather
-        than refer to the holdings directory.
-
-        Keyword arguments:
-            abspath -- the absolute path of a file or a directory.
-        """
-
-        if cls.SHELVES_ONLY:
-            try:
-                (shelf_abspath,
-                 key) = cls.shelf_path_and_key_for_abspath(abspath, 'info')
-
-                if key:
-                    shelf = cls._get_shelf(shelf_abspath,
-                                               log_missing_file=False)
-                    (_, _, _, checksum, _) = shelf[key]
-                    return (checksum == '')
-                elif cls.os_path_exists(shelf_abspath):
-                    return True     # Every shelf file has an entry with an
-                                    # empty key, so this avoids an unnecessary
-                                    # open of the file.
-                else:
-                    return False
-            except (ValueError, IndexError, IOError, OSError):
-                pass
-
-            # Maybe it's associated with something else in the infoshelf tree
-            if f'/{cls.PDS_HOLDINGS}/' in abspath:
-
-                # Maybe there's an associated directory in the infoshelf tree
-                shelf_abspath = abspath.replace(f'/{cls.PDS_HOLDINGS}/',
-                                                f'/{cls.PDS_HOLDINGS}/_infoshelf-')
-                if os.path.exists(shelf_abspath):
-                    return True
-
-                # Maybe there's an associated shelf file in the infoshelf tree
-                if os.path.exists(shelf_abspath + '_info.pickle'):
-                    return True
-
-                # Checksum files need special handling
-                testpath = cls._non_checksum_abspath(abspath)
-                if testpath and cls.os_path_exists(testpath):
-                    # If the testpath exists, then whether it is a directory or
-                    # not depends on the extension
-                    return (not abspath.lower().endswith('.txt'))
-
-        return os.path.isdir(abspath)
-
-    @classmethod
-    def os_listdir(cls, abspath):
-        """Return a list of the file basenames within a directory, given its absolute
-        path. This replaces os.listdir() but might use infoshelf files rather than the
-        file system.
-
-        Keyword arguments:
-            abspath -- the given absolute path.
-        """
-
-        # Make sure there is no trailing slash
-        abspath = abspath.rstrip('/')
-
-        if cls.SHELVES_ONLY:
-            try:
-                (shelf_abspath,
-                 key) = cls.shelf_path_and_key_for_abspath(abspath, 'info')
-
-                shelf = cls._get_shelf(shelf_abspath,
-                                           log_missing_file=False)
-            except (ValueError, IndexError, IOError, OSError):
-                pass
-            else:
-                # Look for paths that begin the same and do not have an
-                # additional slash
-                prefix = key + '/' if key else ''
-                lprefix = len(prefix)
-                basenames = []
-                for key in shelf.keys():
-                    if not key.startswith(prefix): continue
-                    if key == '': continue
-                    basename = key[lprefix:]
-                    if '/' not in basename:
-                        basenames.append(basename)
-
-                return basenames
-
-            # Deal with checksums-archives directories
-            if f'/{cls.PDS_HOLDINGS}/checksums-archives-' in abspath:
-                if abspath.endswith('.txt'):
-                    return []
-
-                testpath = abspath.replace('/checksums-','/')
-                results = cls.os_listdir(testpath)
-
-                for voltype in cls.VOLTYPES:
-                    if '-' + voltype in abspath:
-                        if voltype == 'volumes':
-                            return [r + '_md5.txt' for r in results]
-                        else:
-                            return [r + '_' + voltype + '_md5.txt' for r in results]
-
-                raise ValueError('Invalid abspath for os_listdir: ' + abspath)
-
-            # Deal with checksums directories
-            if f'/{cls.PDS_HOLDINGS}/checksums-' in abspath:
-                if abspath.endswith('_md5.txt'):
-                    return []
-
-                testpath = abspath.replace('/checksums-','/')
-                results = cls.os_listdir(testpath)
-
-                after = abspath.rpartition(f'/{cls.PDS_HOLDINGS}/checksums-')[-1]
-                parts = after.split('/')
-                if len(parts) == 1:         # category-level call
-                    return results
-
-                voltype = parts[0]
-                if voltype == 'volumes' or voltype == 'bundles':
-                    return [r + '_md5.txt' for r in results]
-                else:
-                    return [r + '_' + voltype + '_md5.txt' for r in results]
-
-            # Deal with archive directories
-            if f'/{cls.PDS_HOLDINGS}/archives-' in abspath:
-                if abspath.endswith('.tar.gz'):
-                    return []
-
-                testpath = abspath.replace('/archives-','/')
-                results = cls.os_listdir(testpath)
-
-                after = abspath.rpartition(f'/{cls.PDS_HOLDINGS}/archives-')[-1]
-                parts = after.split('/')
-                if len(parts) == 1:         # category-level call
-                    return results
-
-                voltype = parts[0]
-                if voltype == 'volumes':
-                    return [r + '.tar.gz' for r in results]
-                else:
-                    return [r + '_' + voltype + '.tar.gz' for r in results]
-
-            # Deal with other holdings directories, e.g., holdings/volumes
-            if f'/{cls.PDS_HOLDINGS}/' in abspath:
-
-                # Maybe there's an associated directory in the infoshelf tree
-                shelf_abspath = abspath.replace(f'/{cls.PDS_HOLDINGS}/',
-                                                f'/{cls.PDS_HOLDINGS}/_infoshelf-')
-                try:
-                    results = os.listdir(shelf_abspath)
-                except FileNotFoundError:
-                    # If the shelf file is missing, try the actual file system
-                    # For documentation, we have all files available but not the shelf
-                    # files, therefore we will check the actual file system for documents.
-                    childnames = os.listdir(abspath)
-                    return [c for c in childnames
-                            if c != '.DS_Store' and not c.startswith('._')]
-
-                if not results:
-                    return []
-
-                after = abspath.rpartition(f'/{cls.PDS_HOLDINGS}/')[-1]
-                parts = after.split('/')
-                if len(parts) == 1:         # category-level call
-                    return results
-
-                # Isolate unique bundle names from shelf files
-                # This prevent duplicated results for _info.py and _info.pickle
-                filtered = []
-                for result in results:
-                    parts = result.split('_info.')
-                    if len(parts) == 1: continue
-
-                    bundlename = parts[0]
-                    if bundlename not in filtered:
-                        filtered.append(bundlename)
-
-                # Check the actual file system for a bundleset-level AAREADME
-                aareadmes = []
-                for basename in cls.EXTRA_README_BASENAMES:
-                    if os.path.exists(abspath + '/' + basename):
-                        aareadmes.append(basename)
-
-                return aareadmes + filtered
-
-        childnames = os.listdir(abspath)
-        return [c for c in childnames
-                if c != '.DS_Store' and not c.startswith('._')]
-
-    @classmethod
-    def glob_glob(cls, abspath, force_case_sensitive=False):
-        """Return a list of the existing absolute paths. Works the same as glob.glob(),
-        but uses shelf files instead of accessing the filesystem directly.
-
-        Note: This function is case-insensitive under SHELVES_ONLY. Otherwise,
-        its behavior matches that of the file system. For Macs, this usually
-        means that it is case insensitive. If force_case_sensitive=True, then
-        file paths will only match if the case is exact.
-
-        Keyword arguments:
-            abspath              -- the given absolute path
-            force_case_sensitive -- a flag to determine if the filepath will be case
-                                    sensitive (default False)
-        """
-
-        # We can save a lot of trouble if there's no match pattern
-        # This also enables support for index row notation "index.tab/whatever"
-        if not _needs_glob(abspath):
-            if cls.os_path_exists(abspath, force_case_sensitive):
-                return [abspath]
-            else:
-                return []
-
-        if not cls.SHELVES_ONLY:
-            return _clean_glob(cls, abspath, force_case_sensitive)
-
-        # Find the shelf file(s) if any
-        abspath = abspath.rstrip('/')
-        try:
-            (pattern, key) = cls.shelf_path_and_key_for_abspath(abspath, 'info')
-        except ValueError:
-            # For a category-level holdings dir, this might still work
-            if f'/{cls.PDS_HOLDINGS}/' in abspath:
-                pattern = abspath.replace(f'/{cls.PDS_HOLDINGS}/',
-                                          f'/{cls.PDS_HOLDINGS}/_infoshelf-')
-                key = None  # Below, None indicates that we handled this error
-            else:
-                pattern = ''
-
-        if not pattern:
-            shelf_paths = []
-        elif _needs_glob(pattern):
-            shelf_paths = _clean_glob(cls, pattern)
-        elif os.path.exists(pattern):
-            shelf_paths = [pattern]
-        else:
-            shelf_paths = []
-
-        # If there are no exact infoshelf files, revert to the file system
-        if not shelf_paths:
-            return _clean_glob(cls, abspath, force_case_sensitive)
-
-        # If the check for an exact shelf file failed, just convert the list
-        # of shelf/info directories back to holdings directories
-        if key is None:
-            return [p.replace(f'/{cls.PDS_HOLDINGS}/_infoshelf-', f'/{cls.PDS_HOLDINGS}/')
-                    for p in shelf_paths]
-
-        # Gather the matching entries in each shelf
-        abspaths = []
-        for shelf_path in shelf_paths:
-            shelf = cls._get_shelf(shelf_path)
-            parts = shelf_path.split(f'/{cls.PDS_HOLDINGS}/_infoshelf-')
-            assert len(parts) == 2
-
-            root_ = parts[0] + f'/{cls.PDS_HOLDINGS}/' + parts[1].split('_info.')[0] + '/'
-
-            if _needs_glob(key):
-                # Since shelf files are always in alphabetical order, we can
-                # use a binary search to figure out where to start comparing
-                # strings. This is useful because there can be a lot of
-                # paths to search through, and fnmatchcase is slow.
-                w1 = key.find('?')
-                w2 = key.find('*')
-                w3 = key.find('[')
-                wildcard_index = len(key)
-                if w1 != -1:
-                    wildcard_index = w1
-                if w2 != -1:
-                    wildcard_index = min(wildcard_index, w2)
-                if w3 != -1:
-                    wildcard_index = min(wildcard_index, w3)
-                key_prefix = key[:wildcard_index]
-                interior_paths = list(shelf.keys())
-                values = list(shelf.values())
-                starting_pos = bisect.bisect_left(interior_paths, key_prefix)
-                num_key_slashes = len(key.split('/'))
-                for (interior_path, value) in zip(
-                                interior_paths[starting_pos:],
-                                values[starting_pos:]):
-                    # If the key prefix doesn't match the interior_path prefix,
-                    # then we're done since the filenames are in alphabetical
-                    # order.
-                    if (key_prefix.upper() !=
-                        interior_path[:wildcard_index].upper()):
-                        break
-                    # Because fnmatch matches strings instead of filesystems,
-                    # it has the unfortunate property that match patterns can
-                    # accidentally cross directory boundaries. For example, the
-                    # pattern "f*r" will match "foo/bar", when it shouldn't. We
-                    # handle this by also checking that the returned result
-                    # contains the same number of slashes as the pattern.
-                    if (fnmatch.fnmatchcase(interior_path, key) and
-                        len(interior_path.split('/')) == num_key_slashes):
-                            abspaths.append(root_ + interior_path)
-            else:
-                if key in shelf:
-                    abspaths.append(root_ + key)
-
-        # Remove trailing slashes!
-        return [p.rstrip('/') for p in abspaths]
 
     ############################################################################
     # Properties
@@ -4890,294 +4489,6 @@ class PdsFile(object):
     SHELF_ACCESS_COUNT = 0
 
     SHELF_NULL_KEY_VALUES = {}
-
-    def shelf_path_and_lskip(self, shelf_type='info', bundlename=''):
-        """Return the absolute path to the shelf file associated with this PdsFile.
-        Also return the number of characters to skip over in that absolute path to obtain
-        the key into the shelf.
-
-        Keyword arguments:
-            shelf_type -- shelf type ID: 'index', 'info', or 'link' (default 'info')
-            bundlename -- an optional bundle name to append to the end of a this path,
-                          which can be used if this is a bundleset (default '')
-        """
-
-        if self.checksums_:
-            raise ValueError('No shelf files for checksums: ' +
-                             self.logical_path)
-
-        cls = type(self)
-
-        (dir_prefix, file_suffix) = cls.SHELF_PATH_INFO[shelf_type]
-
-        if self.archives_:
-            if not self.bundleset_:
-                raise ValueError('Archive shelves require bundle sets: ' +
-                                 self.logical_path)
-
-            abspath = ''.join([self.root_, dir_prefix,
-                               self.category_, self.bundleset, self.suffix,
-                               file_suffix, '.pickle'])
-            lskip = (len(self.root_) + len(self.category_) +
-                     len(self.bundleset_))
-
-        else:
-            if bundlename:
-                this_bundlename = bundlename.rstrip('/')
-            else:
-                this_bundlename = self.bundlename
-
-            if not self.bundlename_ and not bundlename:
-                # for non-bundle directories under a bundleset
-                if (self.basename.startswith('checksums_') or
-                    self.basename.startswith('superseded') or
-                    self.basename.endswith('_support')):
-                    this_bundlename = self.basename
-                else:
-                    raise ValueError('Non-archive shelves require bundle names: ' +
-                                     self.logical_path)
-
-            abspath = ''.join([self.root_, dir_prefix,
-                               self.category_, self.bundleset_, this_bundlename,
-                               file_suffix, '.pickle'])
-            lskip = (len(self.root_) + len(self.category_) +
-                     len(self.bundleset_) + len(this_bundlename) + 1)
-
-        return (abspath, lskip)
-
-    def shelf_path_and_key(self, shelf_id='info', bundlename=''):
-        """Return the absolute path to a shelf file, plus the key for this item.
-
-        Keyword arguments:
-            shelf_id   -- shelf type ID: 'index', 'info', or 'link' (default 'info')
-            bundlename -- an optional bundle name to append to the end of a this path,
-                          which can be used if this is a bundleset (default '')
-        """
-
-        (abspath, lskip) = self.shelf_path_and_lskip(shelf_id, bundlename)
-        if bundlename:
-            return (abspath, '')
-        else:
-            return (abspath, self.interior)
-
-    @classmethod
-    def _get_shelf(cls, shelf_path, log_missing_file=True):
-        """Internal method to open and return a shelf/pickle file. A limited number of
-        shelf files are kept open at all times to reduce file IO.
-
-        Use log_missing_file = False to suppress log entries when a nonexistent
-        shelf file is requested but the exception is handled externally.
-
-        Keyword arguments:
-            shelf_path       -- the path of the shelf file
-            log_missing_file -- a flag used to determine if we would like to log the path
-                                of the opening pickle file (default True)
-        """
-
-        # If the shelf is already open, update the access count and return it
-        if shelf_path in cls.SHELF_CACHE:
-            cls.SHELF_ACCESS_COUNT += 1
-            cls.SHELF_ACCESS[shelf_path] = cls.SHELF_ACCESS_COUNT
-
-            return cls.SHELF_CACHE[shelf_path]
-
-        if log_missing_file or os.path.exists(shelf_path):
-            cls.LOGGER.debug('Opening pickle file', shelf_path)
-
-        if not os.path.exists(shelf_path):
-            raise IOError('Pickle file not found: %s' % shelf_path)
-
-        try:
-            with open(shelf_path, 'rb') as f:
-                shelf = pickle.load(f)
-        except Exception as e:
-            raise IOError('Unable to open pickle file: %s' % shelf_path)
-
-        # The pickle files do not produce dictionaries that are in
-        # alphabetical order, so we sort them here in case we want to
-        # do a binary search later.
-        keys_vals = list(zip(shelf.keys(), shelf.values()))
-        keys_vals.sort(key=lambda x: x[0])
-        shelf = dict(keys_vals)
-
-        # Save the null key values from the info shelves. This can save a lot of
-        # shelf open/close operations when we just need info about a bundle,
-        # not an interior file.
-        if '' in shelf and shelf_path not in cls.SHELF_NULL_KEY_VALUES:
-            cls.SHELF_NULL_KEY_VALUES[shelf_path] = shelf['']
-
-        cls.SHELF_ACCESS_COUNT += 1
-        cls.SHELF_ACCESS[shelf_path] = cls.SHELF_ACCESS_COUNT
-        cls.SHELF_CACHE[shelf_path] = shelf
-
-        # Trim the cache if necessary
-        if len(cls.SHELF_CACHE) > (cls.SHELF_CACHE_SIZE +
-                                       cls.SHELF_CACHE_SLOP):
-            pairs = [(cls.SHELF_ACCESS[k],k) for k in cls.SHELF_CACHE]
-            pairs.sort()
-
-            shelf_paths = [p[1] for p in pairs]
-            for shelf_path in shelf_paths[:-cls.SHELF_CACHE_SIZE]:
-                cls._close_shelf(shelf_path)
-
-        return shelf
-
-    @classmethod
-    def _close_shelf(cls, shelf_path):
-        """Internal method to close a shelf file. A limited number of shelf
-        files are kept open at all times to reduce file IO.
-
-        Keyword arguments:
-            shelf_path -- the path of the shelf file
-        """
-
-        # If the shelf is not already open, return
-        if shelf_path not in cls.SHELF_CACHE:
-            cls.LOGGER.error('Cannot close pickle file; not currently open',
-                         shelf_path)
-            return
-
-        # Remove from the cache
-        del cls.SHELF_CACHE[shelf_path]
-        del cls.SHELF_ACCESS[shelf_path]
-
-        cls.LOGGER.debug('Pickle file closed', shelf_path)
-
-    @classmethod
-    def close_all_shelves(cls):
-        """Close all shelf files."""
-
-        keys = list(cls.SHELF_CACHE.keys())     # save keys so dict can be
-        for shelf_path in keys:                     # be modified inside loop!
-            cls._close_shelf(shelf_path)
-
-    def shelf_lookup(self, shelf_type='info', bundlename=''):
-        """Return the contents of a shelf file associated with this object.
-
-        Keyword arguments:
-            shelf_type -- indicates the type of the shelf file: 'info', 'link', or
-                          'index' (default 'info')
-            bundlename -- can be used to get info about a bundle when the method
-                          is applied to its enclosing bundleset (default '')
-        """
-
-        cls = type(self)
-        (shelf_path, key) = self.shelf_path_and_key(shelf_type, bundlename)
-
-        # This potentially saves the need for a lot of opens and closes when
-        # getting info about bundles rather than interior files
-        if key == '':
-            try:
-                return cls.SHELF_NULL_KEY_VALUES[shelf_path]
-            except KeyError:
-                pass
-
-            # Try the second line of the .py file; this is quicker than reading
-            # the whole .pickle file. This is useful because it avoids the need
-            # to open every info shelf file during preload.
-            if shelf_type == 'info':
-                py_path = shelf_path.replace('.pickle', '.py')
-                cls.LOGGER.debug('Retrieving key "%s"' % py_path)
-
-                with open(py_path) as f:
-                    rec = f.readline()
-                    rec = f.readline()
-
-                # Format is "": (bytes, count, date, checksum, (0,0)),
-                parts = rec.partition(':')
-                values = eval(parts[2].strip()[:-1])
-                cls.SHELF_NULL_KEY_VALUES[shelf_path] = values
-                return values
-
-        shelf = cls._get_shelf(shelf_path)
-        return shelf[key]
-
-    @classmethod
-    def shelf_path_and_key_for_abspath(cls, abspath, shelf_type='info'):
-        """Return the absolute path to the shelf file associated with this file path.
-        Also return the key for indexing into the shelf.
-
-        Keyword arguments:
-            abspath    -- the absolute path of the file
-            shelf_type -- shelf type, e.g., 'info' or 'link' (default 'info')
-        """
-
-        # No checksum shelf files allowed
-        (root, _, logical_path) = abspath.partition(f'/{cls.PDS_HOLDINGS}/')
-        if logical_path.startswith('checksums'):
-            raise ValueError('No shelf files for checksums: ' + logical_path)
-
-        (dir_prefix, file_suffix) = cls.SHELF_PATH_INFO[shelf_type]
-
-        # For archive files, the shelf is associated with the bundleset
-        if logical_path.startswith('archives'):
-            parts = logical_path.split('/')
-            if len(parts) < 2:
-                raise ValueError('Archive shelves require bundle sets: ' +
-                                 logical_path)
-
-            shelf_abspath = ''.join([root, f'/{cls.PDS_HOLDINGS}/', dir_prefix,
-                                     parts[0], '/', parts[1],
-                                     file_suffix, '.pickle'])
-            key = '/'.join(parts[2:])
-
-        # Otherwise, the shelf is associated with the bundle
-        else:
-            parts = logical_path.split('/')
-            if len(parts) < 3:
-                raise ValueError('Non-archive shelves require bundle names: ' +
-                                 logical_path)
-
-            shelf_abspath = ''.join([root, f'/{cls.PDS_HOLDINGS}/', dir_prefix,
-                                     parts[0], '/', parts[1], '/', parts[2],
-                                     file_suffix, '.pickle'])
-            key = '/'.join(parts[3:])
-
-        return (shelf_abspath, key)
-
-    @property
-    def info_shelf_expected(self):
-        """Return True if this object should be associated with an entry in an info
-        shelf file.
-        """
-
-        # Checksum files have no info shelves
-        if self.checksums_:
-            return False
-
-        # The document tree does not have info shelves
-        if self.is_documents:
-            return False
-
-        # Category-level directories are merged
-        if self.is_category_dir:
-            return False
-
-        # Archives have info shelves from the bundleset level on down
-        if self.archives_:
-            return True
-
-        # Other files have shelves from the bundlename level on down
-        if self.bundlename:
-            return True
-
-        # This leaves bundleset-level files and their AAREADMEs
-        return False
-
-    def shelf_exists_if_expected(self):
-        """Return True if shelf exists for a pdsfile instance expected to have the shelf
-        file. False if shelf doesn't exist for a pdsfile instance expected to have one.
-        """
-
-        if self.info_shelf_expected:
-            try:
-                self.shelf_lookup('info')
-                return True
-            except OSError:
-                return False
-
-        # Return None if a pdsfile instance doesn't expect the shelf file
-        return None
 
     ############################################################################
     # Log path associations
